remove util mod from sst mod
