+++ analytic_engine/src/sst/mod.rs
@@ -10,4 +10,3 @@ pub mod meta_cache;
 pub mod metrics;
 pub mod parquet;
 pub mod reader;
-pub mod util;
